small refactor of account page and footer
